+++ tests/pyincore_viz/test_pyincore_viz.py
@@ -134,7 +134,7 @@ def test_plot_fragility(client):
     fragility_set = FragilityCurveSet(FragilityService(client).get_dfr3_set("602f31f381bd2c09ad8efcb4"))
     # comment on and off to compare curves
     # plt = plot.get_fragility_plot_2d(fragility_set, title="refactored fragility 2d curve")
-    plt = plot.get_fragility_plot_2d(fragility_set, title="refactored fragility 2d curve", 
+    plt = plot.get_fragility_plot_2d(fragility_set, title="refactored fragility 2d curve",
                                      custom_curve_parameters={"ffe_elev": 3})
     # you can now also plot refactored fragility curve using the main plot method
     # plt = plot.get_fragility_plot(fragility_set, title="refactored fragility 2d curve",
@@ -146,7 +146,7 @@ def test_plot_fragility(client):
     # new format 3d
     fragility_set = FragilityCurveSet(FragilityService(client).get_dfr3_set("5f6ccf67de7b566bb71b202d"))
     plt = plot.get_fragility_plot_3d(fragility_set, title="refactored fragility 3d curve",
-                                                limit_state="LS_0")
+                                     limit_state="LS_0")
     # you can now also plot refactored fragility curve using the main plot method
     # plt = plot.get_fragility_plot(fragility_set, title="refactored fragility 3d curve", limit_state="LS_0",
     #                               dimension=3, custom_curve_parameters={"ffe_elev": 3})